requirements.txt
dgl==2.1.0
fastdtw==0.3.4
gensim==4.3.2
geopy==2.4.1
hyperopt==0.2.7
matplotlib==3.8.2
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
pandas==2.2.2
progressbar33==2.4
ray==2.20.0
scikit_learn==1.3.2
scipy==1.13.0
statsmodels==0.14.2
torch==2.1.2
torchdiffeq==0.2.3
torchtext==0.18.0
tqdm==4.66.1
